refactor: split prompts.py into prompts/ package
Break the monolithic prompts.py into a modular package with one file per
agent (coordinator, modeler, coder, writer) plus shared utilities.
All existing imports remain backward-compatible via __init__.py re-exports.
