+++ backend/app/core/prompts/__init__.py
@@ -1,0 +1,15 @@
+from app.core.prompts.coordinator import COORDINATOR_PROMPT, FORMAT_QUESTIONS_PROMPT
+from app.core.prompts.modeler import MODELER_PROMPT
+from app.core.prompts.coder import CODER_PROMPT
+from app.core.prompts.writer import get_writer_prompt
+from app.core.prompts.shared import get_reflection_prompt, get_completion_check_prompt
+
+__all__ = [
+    "COORDINATOR_PROMPT",
+    "FORMAT_QUESTIONS_PROMPT",
+    "MODELER_PROMPT",
+    "CODER_PROMPT",
+    "get_writer_prompt",
+    "get_reflection_prompt",
+    "get_completion_check_prompt",
+]
